+++ src/cpu/opcodes/sbc.rs
@@ -180,9 +180,9 @@ pub fn sbc_indirect<B: MemoryBus>(cpu: &mut Cpu, bus: &mut B) -> u8 {
 
 fn perform_subtraction_with_carry_u8(cpu: &mut Cpu, value: u16) {
     let old_accumulator_full = cpu.registers.a;
-    let old_accumulator = (old_accumulator_full & 0x00FF) as u16;
+    let old_accumulator = (old_accumulator_full & 0x00FF);
     let v = value & 0x00FF;
-    let carry_in = get_carry_in(cpu) as u16; // 0 or 1
+    let carry_in = get_carry_in(cpu); // 0 or 1
     let diff = (old_accumulator as i16) - (v as i16) - (1 - carry_in as i16);
     let result8 = (diff as u16) & 0x00FF;
